// src/app/api/model/userResponse.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';


export interface UserResponse { 
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    role: Role;
    /**
     * Null bei Super-Admins ohne Organisationszugehörigkeit
     */
    organizationId?: string | null;
    createdAt: string;
}
export namespace UserResponse {
}