// src/app/api/model/organizationResponse.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrganizationResponse { 
    id: string;
    name: string;
    adminEmail: string;
    /**
     * Domain-Whitelist (optional). Wenn gesetzt, können sich Nutzer mit dieser E-Mail-Domain selbst registrieren.
     */
    allowedDomain?: string | null;
    createdAt: string;
}