// src/app/api/model/memberResponse.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';


export interface MemberResponse { 
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    role: Role;
    joinedAt: string;
}
export namespace MemberResponse {
}