// src/app/api/model/taskStatus.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Status eines Tasks: - `OPEN`: Noch nicht begonnen - `IN_PROGRESS`: Wird gerade bearbeitet - `DONE`: Abgeschlossen (setzt `completedAt`) 
 */
export const TaskStatus = {
    Open: 'OPEN',
    InProgress: 'IN_PROGRESS',
    Done: 'DONE'
} as const;
export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];