// src/app/api/model/recurrenceType.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Wiederholungstyp (MVP-Modell, erweiterbar zu RRULE): - `DAILY`: Täglich (z.B. E-Mails checken) - `WEEKLY`: Wöchentlich - `MONTHLY`: Monatlich 
 */
export const RecurrenceType = {
    Daily: 'DAILY',
    Weekly: 'WEEKLY',
    Monthly: 'MONTHLY'
} as const;
export type RecurrenceType = typeof RecurrenceType[keyof typeof RecurrenceType];